Fixed mis-cast in properties editor

+++ src/widgets/properties_editor.cpp
@@ -11,12 +11,12 @@
 
 template<typename Coll1, typename Coll2>
 static float diff(Coll1 const& older, Coll2 const& newer, size_t n) {
-    for (int i = 0; i < n; ++i) {
-        if (older[i] != newer[i]) {
+    for (int i = 0; i < static_cast<int>(n); ++i) {
+        if (static_cast<float>(older[i]) != static_cast<float>(newer[i])) {
             return newer[i];
         }
     }
-    return older[0];
+    return static_cast<float>(older[0]);
 }
 
 bool osmv::Properties_editor::draw(OpenSim::Component& component) {
